requirements.txt
crewai==0.85.0
langgraph
langchain-community
python-dotenv
